0.3.0

+++ depends_core/src/execution/node/node_hash.rs
@@ -33,7 +33,7 @@ pub enum NodeHash {
     #[default]
     NotHashed,
     /// Equal to another value if the internal number is equal.
-    Hashed(usize),
+    Hashed(u64),
 }
 
 impl PartialEq for NodeHash {
